+++ app/src/components/Navbar/index.js
@@ -1,38 +1,36 @@
-import React, { PropTypes } from 'react';
-import Header from 'grommet-udacity/components/Header';
-import Title from 'grommet-udacity/components/Title';
-import Anchor from 'grommet-udacity/components/Anchor';
-import Search from 'grommet-udacity/components/Search';
+// @flow
+import React from 'react';
+import Header from 'grommet/components/Header';
+import Title from 'grommet/components/Title';
+import Anchor from 'grommet/components/Anchor';
+import Search from 'grommet/components/Search';
 import LogoImage from './logo.png';
 import { StyledMenu, StyledLogo, LogoPlaceholder } from './styles';
 
-const Navbar = ({
-  pathname,
-}) => (
-  <div>
-    <Header justify="between">
-      <Title>
-        {typeof window !== 'undefined' ?
-          <StyledLogo src={LogoImage} alt="logo" />
-        :
-          <LogoPlaceholder />
-        }
-      </Title>
-      <StyledMenu inline direction="row" align="center" responsive={false}>
-        <Anchor href="/" className={pathname === '/' ? 'active' : ''}>
-          Home
-        </Anchor>
-        <Anchor href="/about" className={pathname === '/about' ? 'active' : ''}>
-          About
-        </Anchor>
-        <Search dropAlign={{ right: 'right' }} />
-      </StyledMenu>
-    </Header>
-  </div>
-);
-
-Navbar.propTypes = {
-  pathname: PropTypes.string.isRequired,
-};
-
-export default Navbar;
+export default function Navbar(props: {
+  pathname: string
+}) {
+  const { pathname } = props;
+  return (
+    <div>
+      <Header justify="between">
+        <Title>
+          {typeof window !== 'undefined' ?
+            <StyledLogo src={LogoImage} alt="logo" />
+          :
+            <LogoPlaceholder />
+          }
+        </Title>
+        <StyledMenu inline direction="row" align="center" responsive={false}>
+          <Anchor path="/" className={pathname === '/' ? 'active' : ''}>
+            Home
+          </Anchor>
+          <Anchor path="/about" className={pathname === '/about' ? 'active' : ''}>
+            About
+          </Anchor>
+          <Search dropAlign={{ right: 'right' }} />
+        </StyledMenu>
+      </Header>
+    </div>
+  );
+}