+++ app/src/containers/AppContainer/index.js
@@ -1,9 +1,9 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import App from 'grommet-udacity/components/App';
-import { Navbar, AppFooter } from 'components';
+import App from 'grommet/components/App';
 import { AppContainer as ReactHotLoader } from 'react-hot-loader';
+import { Navbar, AppFooter } from 'components';
 import * as AppContainerActionCreators from './actions';
 
 class AppContainer extends Component { // eslint-disable-line react/prefer-stateless-function