+++ server/app.js
@@ -1,7 +1,9 @@
-/* eslint-disable */
+/* eslint-disable no-console */
 import express from 'express';
+import path from 'path';
 import morgan from 'morgan';
 import React from 'react';
+import env from 'node-env-file';
 import { renderToString, renderToStaticMarkup } from 'react-dom/server';
 import { match, RouterContext } from 'react-router';
 import { Provider } from 'react-redux';
@@ -11,14 +13,17 @@ import { routes } from '../app/src/routes';
 import Html from './utils/Html';
 import manifest from './public/manifest.json';
 
+env(path.join(__dirname, '..', '.env'));
+
 const app = express();
+const serverUrl = process.env.BASE_URL || 'http://localhost:1337';
+const apiUrl = process.env.API_URL || 'http://localhost:3000';
+const PORT = serverUrl.match(/\d+/g)[0];
+const IP = serverUrl.match(/\w+/g)[1];
+const debug = process.env.DEBUG === 'true';
 
-// Need to set this to your api url
-const IP = process.env.IP || '0.0.0.0';
-const PORT = process.env.PORT || 1337;
-
-app.use(morgan('combined'));
-app.use(express.static(__dirname + '/public'));
+if (debug) { app.use(morgan('combined')); }
+app.use(express.static(path.join(__dirname, '/public')));
 
 app.use((req, res) => {
   match({ routes, location: req.url },
@@ -59,4 +64,3 @@ app.listen(PORT, IP, (err) => {
   }
   return console.info(`==> 😎 Listening on port ${PORT}. Open http://${IP}:${PORT} in your browser.`);
 });
-/* eslint-enable */